Conflicts:
	hdr/dataRT.h
	hdr/miniRT.h

	modified:   README.md
	modified:   hdr/dataRT.h
	modified:   hdr/miniRT.h
	renamed:    scene.rt -> model.rt
	modified:   render.png
	modified:   src/ft_deadzn.c
	modified:   src/ft_det3x3.c
	modified:   src/ft_distce.c
	modified:   src/ft_invmat.c
	modified:   src/ft_keyhook.c
	modified:   src/ft_multlv.c
	modified:   src/ft_newcyl.c
	modified:   src/ft_newlig.c
	modified:   src/ft_newsph.c
	modified:   src/ft_norm.c
	modified:   src/ft_nrmlze.c
	modified:   src/ft_pars_coord_rgb.c
	modified:   src/ft_polyd2.c
	modified:   src/ft_raytra.c
	modified:   src/ft_render.c
	modified:   src/ft_scalar.c
	modified:   src/ft_setamb.c
	modified:   src/ft_setcam.c
	modified:   src/ft_setvec.c
	modified:   src/ft_shades.c
	modified:   src/ft_shdcyl.c
	modified:   src/ft_shdpla.c
	modified:   src/ft_shdsph.c
	modified:   src/ft_syscyl.c
	modified:   src/ft_syspla.c
	modified:   src/ft_syssph.c

+++ src/ft_deadzn.c
@@ -6,13 +6,13 @@
 /*   By: alfux <alexis.t.fuchs@gmail.com>           +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/12/23 11:52:51 by alfux             #+#    #+#             */
-/*   Updated: 2022/12/23 11:59:26 by alfux            ###   ########.fr       */
+/*   Updated: 2022/12/28 23:59:18 by alfux            ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include <miniRT.h>
 
-int	ft_deadzn(t_2x3 *i, t_vec const *p, float dz)
+int	ft_deadzn(t_2x3 *i, t_vec const *p, double dz)
 {
 	if (ft_distce(i->top, *p) <= dz)
 	{
