+++ src/ft_polyd2.c
@@ -6,14 +6,14 @@
 /*   By: alfux <alexis.t.fuchs@gmail.com>           +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/12/06 13:25:09 by alfux             #+#    #+#             */
-/*   Updated: 2022/12/22 15:09:36 by alfux            ###   ########.fr       */
+/*   Updated: 2022/12/29 00:03:44 by alfux            ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 #include <miniRT.h>
 
 t_vec	ft_polyd2(t_vec coef)
 {
-	float	delta;
+	double	delta;
 
 	delta = (coef.y * coef.y) - (4 * coef.x * coef.z);
 	if (delta < 0.f || coef.x == 0.f)
