+++ src/ft_shdpla.c
@@ -6,13 +6,13 @@
 /*   By: alfux <alexis.t.fuchs@gmail.com>           +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/12/23 10:06:25 by alfux             #+#    #+#             */
-/*   Updated: 2022/12/23 10:58:14 by alfux            ###   ########.fr       */
+/*   Updated: 2022/12/29 00:11:50 by alfux            ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include <miniRT.h>
 
-float	ft_shdpla(t_win const *win, t_pla const *pl, t_vec const *p)
+double	ft_shdpla(t_win const *win, t_pla const *pl, t_vec const *p)
 {
 	t_vec	*lpos;
 	t_vec	n;
