+++ src/ft_shdsph.c
@@ -6,13 +6,13 @@
 /*   By: alfux <alexis.t.fuchs@gmail.com>           +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/12/23 09:32:26 by alfux             #+#    #+#             */
-/*   Updated: 2022/12/23 10:57:48 by alfux            ###   ########.fr       */
+/*   Updated: 2022/12/29 00:11:32 by alfux            ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include <miniRT.h>
 
-float	ft_shdsph(t_win const *win, t_sph const *s, t_vec const *p)
+double	ft_shdsph(t_win const *win, t_sph const *s, t_vec const *p)
 {
 	t_vec	*lpos;
 	t_vec	n;
